Start to rework models
