+++ src/app/components/like/like.component.ts
@@ -1,7 +1,7 @@
 import {Component, Input, OnInit} from '@angular/core';
-import {Post} from '../../models/Post';
-import {User} from '../../models/User';
-import {Rating} from '../../models/Rating';
+import {Post} from '../../models/common/Post';
+import {User} from '../../models/common/User';
+import {Rating} from '../../models/common/Rating';
 import {UserService} from '../../services/UserService';
 import {RatingService} from '../../services/RatingService';
 import {Store} from '@ngrx/store';
