--- src/app/models/ImageNode.ts
@@ -1,9 +1,0 @@
-export class Image {
-  id: string;
-  imageUrl: string;
-
-  constructor(id: string, imageUrl: string) {
-    this.id = id;
-    this.imageUrl = imageUrl;
-  }
-}
