--- src/app/models/ListNode.ts
@@ -1,7 +1,0 @@
-export class ListNode {
-  content: string[];
-
-  constructor(content: string[]) {
-    this.content = content;
-  }
-}
