+++ src/app/models/common/ImageNode.ts
@@ -1,0 +1,9 @@
+export class Image {
+  id: string;
+  imageUrl: string;
+
+  constructor(id: string, imageUrl: string) {
+    this.id = id;
+    this.imageUrl = imageUrl;
+  }
+}
