+++ src/app/models/common/ListNode.ts
@@ -1,0 +1,7 @@
+export class ListNode {
+  content: string[];
+
+  constructor(content: string[]) {
+    this.content = content;
+  }
+}
