+++ src/app/models/common/NodeContent.ts
