+++ src/app/models/common/TextNode.ts
