+++ src/app/models/common/VideoNode.ts
