+++ src/app/redux/models/ReduxUser.ts
@@ -1,4 +1,4 @@
-import {User} from '../../models/User';
+import {User} from '../../models/common/User';
 
 export class ReduxUser {
